# docs/openapi-v2.yaml
openapi: 3.1.0
info:
  title: Vook Frontend API
  version: 2.0.0
  description: Resource-oriented contract consumed by the standalone Vook frontend.
servers: [{ url: /api/v2 }]
security: [{ sessionCookie: [] }]
paths:
...
  /expenses/mine:
    get: { operationId: listMyExpenses, responses: { '200': { $ref: '#/components/responses/Object' } } }
    post: { operationId: createExpense, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '201': { $ref: '#/components/responses/Object' } } }
  /expenses/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch: { operationId: updateExpense, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '200': { $ref: '#/components/responses/Object' } } }
  /files/receipts:
    post: { operationId: uploadReceipt, requestBody: { $ref: '#/components/requestBodies/Multipart' }, responses: { '201': { $ref: '#/components/responses/Object' } } }
  /files/documents:
    post: { operationId: uploadDocument, requestBody: { $ref: '#/components/requestBodies/Multipart' }, responses: { '201': { $ref: '#/components/responses/Object' } } }
  /files/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get: { operationId: downloadFile, responses: { '200': { $ref: '#/components/responses/Binary' } } }
  /documents:
    get: { operationId: listDocuments, responses: { '200': { $ref: '#/components/responses/Object' } } }
    post: { operationId: createDocument, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '201': { $ref: '#/components/responses/Object' } } }
...
  /documents/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    delete: { operationId: deleteDocument, responses: { '200': { $ref: '#/components/responses/Empty' } } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: vook_session }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
...
  requestBodies:
    Object: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
    Multipart: { required: true, content: { multipart/form-data: { schema: { type: object, properties: { file: { type: string, format: binary } }, required: [file] } } } }
  responses:
    Empty: { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/EmptyEnvelope' } } } }
    Object: { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/ObjectEnvelope' } } } }
...
    Binary: { description: Download, headers: { Content-Disposition: { schema: { type: string } } }, content: { application/octet-stream: { schema: { type: string, format: binary } } } }
  schemas:
...
    ListMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties: { page: { type: integer, minimum: 1 }, pageSize: { type: integer, minimum: 1 }, total: { type: integer, minimum: 0 }, totalPages: { type: integer, minimum: 0 } }
...
    ObjectEnvelope: { type: object, required: [data], properties: { data: {}, meta: { $ref: '#/components/schemas/ListMeta' } } }
    EmptyEnvelope: { type: object, required: [data], properties: { data: { type: 'null' } } }
...